sync commit

+++ hdl_matmul.m
@@ -34,7 +34,7 @@ function [c_out, state_store] =  hdl_matmul (a_in,b_in)
         x_in = 1;
     end
 
-    c_out = uint32(0) ;
+    c_out = int8(0) ;
     if (0 == state)
         A(x_in, y_in) = a_in;
         B(x_in, y_in) = b_in;
@@ -52,7 +52,7 @@ function [c_out, state_store] =  hdl_matmul (a_in,b_in)
     end
 
     if (1 == state)
-        %C(x_in,y_in) =uint32(double( A(x_in,:)) *double( B (:,y_in)));
+        %C(x_in,y_in) =int8(double( A(x_in,:)) *double( B (:,y_in)));
         C(x_in,y_in) =(( A(x_in,:)) *( B (:,y_in)));
         if (MAT_SIZE == y_in)
             y_in = 1;
@@ -68,7 +68,7 @@ function [c_out, state_store] =  hdl_matmul (a_in,b_in)
     end
 
     if (2 == state)
-        c_out = uint32(C(x_in,y_in));
+        c_out = int8(C(x_in,y_in));
         if (MAT_SIZE == y_in)
             y_in = 1;
             if (MAT_SIZE == x_in)
@@ -81,6 +81,6 @@ function [c_out, state_store] =  hdl_matmul (a_in,b_in)
             y_in = y_in + 1;
         end
     end
-    state_store= uint32(state);
+    state_store= int8(state);
 
 end
